Why did I ever think this was a good idea lol
